[Selection] extract common methods to functions
