+++ src/packages/selection/selection-controller/ClickController.ts
@@ -1,23 +1,13 @@
 // SPDX-FileCopyrightText: 2023 Open Pioneer project (https://github.com/open-pioneer)
 // SPDX-License-Identifier: Apache-2.0
-import { Resource } from "@open-pioneer/core";
 import { MapBrowserEvent } from "ol";
 import OlMap from "ol/Map";
 import { unByKey } from "ol/Observable";
-import Overlay from "ol/Overlay";
 import Geometry from "ol/geom/Geometry";
 import { Polygon } from "ol/geom";
+import { activateViewportInteraction, createHelpTooltip, deactivateViewportInteraction, Tooltip } from "./helper";
+import { EventsKey } from "ol/events";
 
-
-/** Represents a tooltip rendered on the OpenLayers map. */
-interface Tooltip extends Resource {
-    overlay: Overlay;
-    element: HTMLDivElement;
-    setText(value: string): void;
-}
-
-const ACTIVE_CLASS = "selection-active";
-const INACTIVE_CLASS = "selection-inactive";
 
 export class ClickController {
     private tooltip: Tooltip;
@@ -27,20 +17,21 @@ export class ClickController {
     private tooltipDisabledMessage: string;
 
     private onClick: (evt: MapBrowserEvent<UIEvent>) => void;
+    private eventKey: EventsKey;
 
     constructor(
         olMap: OlMap,
         tooltipMessage: string,
         tooltipDisabledMessage: string,
         onExtentSelected: (geometry: Geometry) => void
-    ) {        
-        this.tooltip = this.createHelpTooltip(olMap, tooltipMessage);
+    ) {
+        this.tooltip = createHelpTooltip(olMap, tooltipMessage);
         this.olMap = olMap;
         this.tooltipMessage = tooltipMessage;
         this.tooltipDisabledMessage = tooltipDisabledMessage;
         this.onClick = this.getEventHandlerFunction(onExtentSelected);
-        this.olMap.on("singleclick", this.onClick);
-        this.olMap.getViewport().classList.add(ACTIVE_CLASS);
+        this.eventKey = this.olMap.on("singleclick", this.onClick);
+        activateViewportInteraction(olMap);
     }
 
     /**
@@ -48,25 +39,21 @@ export class ClickController {
      */
     destroy() {
         this.tooltip.destroy();
-        this.olMap.un("singleclick", this.onClick);
-        this.olMap.getViewport().classList.remove(ACTIVE_CLASS);
-        this.olMap.getViewport().classList.remove(INACTIVE_CLASS);
+        unByKey(this.eventKey);
+        deactivateViewportInteraction(this.olMap);
 
     }
 
     setActive(isActive: boolean) {
         if (this.isActive === isActive) return;
-        const viewPort = this.olMap.getViewport();
         if (isActive) {
             this.tooltip.setText(this.tooltipMessage);
-            viewPort.classList.remove(INACTIVE_CLASS);
-            viewPort.classList.add(ACTIVE_CLASS);
+            activateViewportInteraction(this.olMap);
             this.isActive = true;
         } else {
-            this.olMap.un("singleclick", this.onClick);
+            unByKey(this.eventKey);
             this.tooltip.setText(this.tooltipDisabledMessage);
-            viewPort.classList.remove(ACTIVE_CLASS);
-            viewPort.classList.add(INACTIVE_CLASS);
+            deactivateViewportInteraction(this.olMap);
             this.isActive = false;
         }
     }
@@ -78,8 +65,8 @@ export class ClickController {
             const coordinates = evt.coordinate;
             const pixel = map.getPixelFromCoordinate(coordinates);
             const clickTolerance = [
-                [pixel[0]! - pixelTolerance, pixel[1]! - pixelTolerance], 
-                [pixel[0]! - pixelTolerance, pixel[1]! + pixelTolerance], 
+                [pixel[0]! - pixelTolerance, pixel[1]! - pixelTolerance],
+                [pixel[0]! - pixelTolerance, pixel[1]! + pixelTolerance],
                 [pixel[0]! + pixelTolerance, pixel[1]! + pixelTolerance],
                 [pixel[0]! + pixelTolerance, pixel[1]! - pixelTolerance],
                 [pixel[0]! - pixelTolerance, pixel[1]! - pixelTolerance],
@@ -90,44 +77,4 @@ export class ClickController {
         };
         return getExtentFromEvent;
     }
-
-
-
-    /**
-     * Method to generate a tooltip on the mouse cursor
-     */
-    private createHelpTooltip(olMap: OlMap, message: string): Tooltip {
-        const element = document.createElement("div");
-        element.className = "selection-tooltip printing-hide";
-        element.role = "tooltip";
-
-        const content = document.createElement("span");
-        content.textContent = message;
-        element.appendChild(content);
-
-        const overlay = new Overlay({
-            element: element,
-            offset: [15, 0],
-            positioning: "center-left"
-        });
-
-        const pointHandler = olMap.on("pointermove", (evt) => {
-            overlay.setPosition(evt.coordinate);
-        });
-
-        olMap.addOverlay(overlay);
-        return {
-            overlay,
-            element,
-            destroy() {
-                olMap.removeOverlay(overlay);
-                overlay.dispose();
-                unByKey(pointHandler);
-            },
-            setText(value) {
-                content.textContent = value;
-            }
-        };
-    }
-
 }
