+++ src/packages/selection/selection-controller/DragController.ts
@@ -2,25 +2,15 @@
 // SPDX-License-Identifier: Apache-2.0
 import { Resource } from "@open-pioneer/core";
 import OlMap from "ol/Map";
-import { unByKey } from "ol/Observable";
-import Overlay from "ol/Overlay";
 import { mouseActionButton } from "ol/events/condition";
 import Geometry from "ol/geom/Geometry";
 import { DragBox, DragPan } from "ol/interaction";
 import PointerInteraction from "ol/interaction/Pointer";
+import { activateViewportInteraction, createHelpTooltip, deactivateViewportInteraction, Tooltip } from "./helper";
 
 interface InteractionResource extends Resource {
     interaction: PointerInteraction;
 }
-/** Represents a tooltip rendered on the OpenLayers map. */
-interface Tooltip extends Resource {
-    overlay: Overlay;
-    element: HTMLDivElement;
-    setText(value: string): void;
-}
-
-const ACTIVE_CLASS = "selection-active";
-const INACTIVE_CLASS = "selection-inactive";
 
 export class DragController {
     private tooltip: Tooltip;
@@ -42,15 +32,15 @@ export class DragController {
         );
         this.interactionResources.push(this.createDrag(olMap, viewPort, this.interactionResources));
 
-        this.tooltip = this.createHelpTooltip(olMap, tooltipMessage);
+        this.tooltip = createHelpTooltip(olMap, tooltipMessage);
         this.olMap = olMap;
         this.tooltipMessage = tooltipMessage;
         this.tooltipDisabledMessage = tooltipDisabledMessage;
     }
 
     initViewport(olMap: OlMap) {
+        activateViewportInteraction(olMap);
         const viewPort = olMap.getViewport();
-        viewPort.classList.add(ACTIVE_CLASS);
 
         viewPort.oncontextmenu = (e) => {
             e.preventDefault();
@@ -71,22 +61,19 @@ export class DragController {
 
     setActive(isActive: boolean) {
         if (this.isActive === isActive) return;
-        const viewPort = this.olMap.getViewport();
         if (isActive) {
             this.interactionResources.forEach((interaction) =>
                 this.olMap.addInteraction(interaction.interaction)
             );
             this.tooltip.setText(this.tooltipMessage);
-            viewPort.classList.remove(INACTIVE_CLASS);
-            viewPort.classList.add(ACTIVE_CLASS);
+            activateViewportInteraction(this.olMap);
             this.isActive = true;
         } else {
             this.interactionResources.forEach((interaction) =>
                 this.olMap.removeInteraction(interaction.interaction)
             );
             this.tooltip.setText(this.tooltipDisabledMessage);
-            viewPort.classList.remove(ACTIVE_CLASS);
-            viewPort.classList.add(INACTIVE_CLASS);
+            deactivateViewportInteraction(this.olMap);
             this.isActive = false;
         }
     }
@@ -116,8 +103,7 @@ export class DragController {
                 olMap.removeInteraction(dragBox);
                 interactionResources.splice(interactionResources.indexOf(this));
                 dragBox.dispose();
-                viewPort.classList.remove(ACTIVE_CLASS);
-                viewPort.classList.remove(INACTIVE_CLASS);
+                deactivateViewportInteraction(olMap);
                 viewPort.oncontextmenu = null;
             }
         };
@@ -151,8 +137,7 @@ export class DragController {
                 olMap.removeInteraction(drag);
                 interactionResources.splice(interactionResources.indexOf(this));
                 drag.dispose();
-                viewPort.classList.remove(ACTIVE_CLASS);
-                viewPort.classList.remove(INACTIVE_CLASS);
+                deactivateViewportInteraction(olMap);
                 viewPort.oncontextmenu = null;
             }
         };
@@ -160,42 +145,6 @@ export class DragController {
         return interactionResource;
     }
 
-    /**
-     * Method to generate a tooltip on the mouse cursor
-     */
-    private createHelpTooltip(olMap: OlMap, message: string): Tooltip {
-        const element = document.createElement("div");
-        element.className = "selection-tooltip printing-hide";
-        element.role = "tooltip";
-
-        const content = document.createElement("span");
-        content.textContent = message;
-        element.appendChild(content);
-
-        const overlay = new Overlay({
-            element: element,
-            offset: [15, 0],
-            positioning: "center-left"
-        });
-
-        const pointHandler = olMap.on("pointermove", (evt) => {
-            overlay.setPosition(evt.coordinate);
-        });
-
-        olMap.addOverlay(overlay);
-        return {
-            overlay,
-            element,
-            destroy() {
-                olMap.removeOverlay(overlay);
-                overlay.dispose();
-                unByKey(pointHandler);
-            },
-            setText(value) {
-                content.textContent = value;
-            }
-        };
-    }
 
     /**
      * Method for testing purposes only
